feat(Testing): add movement simulator
For many trivial tests one still has to put on the headset in order to
move to the right spot. This increases turn-around times. This script
can be put onto the camera rig and will then allow movement through
the scene using keys. Once at the desired location the controllers can
be used to interact while looking at the monitor.

Supported so far are move, strafe, rotate and up/down. Being an early
version this is only tested in one context so far and especially after
rotating the move directions are inverted currently. These are all
things to look at in future versions. Help always welcome.