chore(Examples): set example scenes up correctly per unity version
As some of the features in SteamVR changed between Unity 5.3 and
Unity 5.4 it means that the example scenes that use the
Headset Collision Fade script didn't work in Unity 5.4 as they were
set up for the SteamVR usage in Unity 5.3.

This commit adds a helper script that adds the Headset Collision Fade
script to the correct game object dependant on the version at
edit time. This is better than doing it at runtime as then users can
view the game objects in the editor without playing the scene and
see where the scripts are supposed to go.

There are also setup scripts for example scene 031 and 034 that set
up the relevant scripts on the Camera on launch of editor.