fix(Interaction): prevent controller swap breaking use functionality
Previously, when using an object and swapping controllers (either
grab or touch state) the use state on the object would not be reset
correctly and therefore the previous using controller would still
be considered the active use controller.

This fix ensures that when a controller stops touching or grabbing
an object, that the use state on the controller is also reset to
no longer using the object.